src/backend/config/agent-templates/professor-polygon/pyproject.toml
[project]
name = "professor-polygon"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "praxis-professor-core @ file:///Users/andrii/Desktop/projects/vybe/agents/praxis-professor-core",
    "pyyaml>=6.0",
    "pandas>=2.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "mypy>=1.11",
    "ruff>=0.6",
    "types-pyyaml>=6.0.12.20260408",
    "pandas-stubs>=3.0.0.260204",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.mypy]
strict = true
python_version = "3.12"
